requirements.txt
flask
google-api-python-client
python-dotenv
nltk
numpy
scikit-learn
vaderSentiment
matplotlib
wordcloud